% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot}
\alias{biplot}
\alias{biplot.pca}
\title{biplot methods for \code{pca} family}
\usage{
\method{biplot}{pca}(
  x,
  comp = c(1, 2),
  ind.names = TRUE,
  group = NULL,
  cutoff = 0,
  col.per.group = NULL,
  col = NULL,
  ind.names.size = 3,
  ind.names.col = color.mixo(4),
  ind.names.repel = TRUE,
  pch = 19,
  pch.levels = NULL,
  pch.size = 2,
  var.names = TRUE,
  var.names.col = "grey40",
  var.names.size = 4,
  var.names.angle = FALSE,
  var.arrow.col = "grey40",
  var.arrow.size = 0.5,
  var.arrow.length = 0.2,
  ind.legend.title = NULL,
  vline = FALSE,
  hline = FALSE,
  legend = if (is.null(group)) FALSE else TRUE,
  legend.title = NULL,
  pch.legend.title = NULL,
  ...
)
}
\arguments{
\item{x}{A 'pca' object.}

\item{comp}{integer vector of length two (or three to 3d). The components
that will be used on the horizontal and the vertical axis respectively to
project the individuals.}

\item{ind.names}{either a character vector of names for the individuals to
be plotted, or \code{FALSE} for no names. If \code{TRUE}, the row names of
the first (or second) data matrix is used as names (see Details).}

\item{group}{Factor indicating the group membership for each sample.}

\item{cutoff}{numeric between 0 and 1. Variables with correlations below
this cutoff in absolute value are not plotted (see Details).}

\item{col.per.group}{character (or symbol) color to be used when 'group' is
defined. Vector of the same length as the number of groups.}

\item{col}{character (or symbol) color to be used, possibly vector.}

\item{ind.names.size}{Numeric, sample name size.}

\item{ind.names.col}{Character, sample name colour.}

\item{ind.names.repel}{Logical, whether to repel away label names.}

\item{pch}{plot character. A character string or a vector of single
characters or integers. See \code{\link{points}} for all alternatives.}

\item{pch.levels}{If \code{pch} is a factor, a named vector providing the
point characters to use. See examples.}

\item{pch.size}{Numeric, sample point character size.}

\item{var.names}{Logical indicating whether to show variable names. 
Alternatively, a character.}

\item{var.names.col}{Character, variable name colour.}

\item{var.names.size}{Numeric, variable name size.}

\item{var.names.angle}{Logical, whether to align variable names to arrow
directions.}

\item{var.arrow.col}{Character, variable arrow colour. If 'NULL', no arrows
are shown.}

\item{var.arrow.size}{Numeric, variable arrow head size.}

\item{var.arrow.length}{Numeric, length of the arrow head in 'cm'.}

\item{ind.legend.title}{Character, title of the legend.}

\item{vline}{Logical, whether to draw the vertical neutral line.}

\item{hline}{Logical, whether to draw the horizontal neutral line.}

\item{legend}{Logical, whether to show the legend if \code{group != NULL}.}

\item{legend.title}{Character, the legend title if \code{group != NULL}.}

\item{pch.legend.title}{Character, the legend title if \code{pch} is a factor.}

\item{...}{Not currently used.}

\item{pch.legend}{Character, the legend title if \code{pch} is a factor.}
}
\value{
A ggplot object.
}
\description{
biplot methods for \code{pca} family
}
\details{
\code{biplot} unifies the reduced representation of both the
observations/samples and variables of a matrix of multivariate data on the
same plot. Essentially, in the reduced space the samples are shown as
points/names and the contributions of features to each dimension are shown as
directed arrows or vectors.
}
\examples{
data("nutrimouse")
pca.lipid <- pca(nutrimouse$lipid, ncomp = 2, scale = TRUE)
# seed for reproducible geom_text_repel
set.seed(42)
biplot(pca.lipid)
## correlation cutoff to filter features
biplot(pca.lipid, cutoff = c(0.8))
## tailor threshold for each component
biplot(pca.lipid, cutoff = c(0.8, 0.7))

## cutomise ggplot in an arbitrary way
biplot(pca.lipid) + theme_linedraw() + 
    # add vline
    geom_vline(xintercept = 0, col = 'green') +
    # add hline
    geom_hline(yintercept = 0, col = 'green') +
    # customise labs
    labs(x = 'Principal Component 1', y = 'Principal Component 2')

## group samples
biplot(pca.lipid, group = nutrimouse$diet, legend.title = 'Diet')

## customise variable labels
biplot(pca.lipid, 
       var.names.col = color.mixo(2),
       var.names.size = 4,
       var.names.angle = TRUE
)

## no arrows
biplot(pca.lipid, group = nutrimouse$diet, legend.title = 'Diet', 
       var.arrow.col = NULL, var.names.col = 'black')

## add x=0 and y=0 lines in function
biplot(pca.lipid, group = nutrimouse$diet, legend.title = 'Diet', 
       var.arrow.col = NULL, var.names.col = 'black', 
       vline = TRUE, hline = TRUE)

## example with spca
spca.lipid <- spca(nutrimouse$lipid, ncomp = 2, scale = TRUE, keepX = c(8, 6))
biplot(spca.lipid, var.names.col = 'black', group = nutrimouse$diet, 
       legend.title = 'Diet')
}
