% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.mproduct.R
\name{dctii_m_transforms}
\alias{dctii_m_transforms}
\title{Create m-transform functions to apply the Discrete Cosine Transform 2}
\usage{
dctii_m_transforms(t, bpparam = NULL)
}
\arguments{
\item{t}{The length of the transform.}

\item{bpparam}{A \linkS4class{BiocParallelParam} object indicating the type
of parallelisation.}
}
\value{
\item{m}{A function which applies the dct-ii along the last dimension of a
given numerical input array. For 3D tensors it performs the mode-3 product
with the DCT matrix.}
\item{minv}{The inverse of m,}
}
\description{
Returns functions \code{m} and \code{m_inv} which apply tubal transforms
defined by the Discrete Cosine Transform (DCT-II variant). This is equivalent
to Scipys DCTI-ii algorithm with \code{norm='ortho'}.
}
\author{
Brendan Lu
}
