% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.mproduct.R
\name{.apply_mat_transform}
\alias{.apply_mat_transform}
\title{Apply a function across the last dimension of an input vector,
matrix, or tensor. This function defines both a parallel algorithm using
BiocParrallel, and also a simple \code{apply} algorithm. Even on Windows
Machines, setting \code{bpparam = BiocParallel::SerialParam()} offers a
notable speedup for larger 3D array inputs.}
\usage{
.apply_mat_transform(x, mat, bpparam)
}
\arguments{
\item{x}{Numerical array input.}

\item{mat}{Function which defines the tubal transform.}

\item{bpparam}{A \linkS4class{BiocParallelParam} object indicating the type
of parallelisation.}
}
\value{
A tensor of the same size under the specified tubal transform,
denoted \hat{x}.
}
\description{
Apply a function across the last dimension of an input vector,
matrix, or tensor. This function defines both a parallel algorithm using
BiocParrallel, and also a simple \code{apply} algorithm. Even on Windows
Machines, setting \code{bpparam = BiocParallel::SerialParam()} offers a
notable speedup for larger 3D array inputs.
}
\author{
Brendan Lu
}
\keyword{internal}
