% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.mproduct.R
\name{.binary_facewise}
\alias{.binary_facewise}
\title{Compute Kilmer's facewise product. Note that the for-loop implementation is
relatively fast, and very readable. There's also a BiocParralel
implementation here, but it lacks significant benchmarking results.
bltodo: the parallel algorithm is probably stupid remove sometime?}
\usage{
.binary_facewise(a, b, bpparam)
}
\description{
Compute Kilmer's facewise product. Note that the for-loop implementation is
relatively fast, and very readable. There's also a BiocParralel
implementation here, but it lacks significant benchmarking results.
bltodo: the parallel algorithm is probably stupid remove sometime?
}
\author{
Brendan Lu
}
\keyword{internal}
