% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpca.R
\name{.extract_tensor_columns}
\alias{.extract_tensor_columns}
\title{Extract tensor columns specified by .unravel_index output.
Effectively achieves:}
\usage{
.extract_tensor_columns(tensor, k_t_indices)
}
\description{
self.loadings_matrix_ = hatV[
        :, self._k_t_flatten_sort[0], self._k_t_flatten_sort[1]
    ].T
}
\details{
type of indexing in Numpy.

This is the 'tensor compression' based on the ordered indices specified
by each column of k_t_indices as described in Mor's TCAM.
}
\author{
Brendan Lu
}
\keyword{internal}
