% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpls.R
\name{.obtain_k_t_flatten_sort}
\alias{.obtain_k_t_flatten_sort}
\title{Convert a singular values tensor (in compressed matrix form) to a set of
indices corresponding to the (column,face) pairs of the top `ncomp` singular
values. NEEDS singular values to be in matrix form.}
\usage{
.obtain_k_t_flatten_sort(s_mat, ncomp)
}
\description{
Convert a singular values tensor (in compressed matrix form) to a set of
indices corresponding to the (column,face) pairs of the top `ncomp` singular
values. NEEDS singular values to be in matrix form.
}
\author{
Brendan Lu
}
\keyword{internal}
