% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpls.R
\name{.obtain_k_t_top}
\alias{.obtain_k_t_top}
\title{Get the k, t index corresponding to the largest singular value. Mirrors
.unravel_index() but more efficient as we only care about the largest
singular value in tpls.}
\usage{
.obtain_k_t_top(s_mat)
}
\description{
Get the k, t index corresponding to the largest singular value. Mirrors
.unravel_index() but more efficient as we only care about the largest
singular value in tpls.
}
\author{
Brendan Lu
}
\keyword{internal}
