% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpca.R
\name{.singular_vals_mat_to_tens}
\alias{.singular_vals_mat_to_tens}
\title{Helper function to convert compressed matrix form of the singular values into
a sparse tensor.}
\usage{
.singular_vals_mat_to_tens(mat, dim)
}
\arguments{
\item{mat}{Matrix s.t. each column contains the f-diagonal singular values}

\item{dim}{Dimension of output tensor}
}
\description{
Helper function to convert compressed matrix form of the singular values into
a sparse tensor.
}
\author{
Brendan Lu
}
\keyword{internal}
