% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.mproduct.R
\name{.stop_invalid_transform_input}
\alias{.stop_invalid_transform_input}
\title{Validate appropriate 'null-ness' of m, minv inputs, and apply default
transform as needed.}
\usage{
.stop_invalid_transform_input(m, minv, t, bpparam)
}
\description{
Validate appropriate 'null-ness' of m, minv inputs, and apply default
transform as needed.
}
\author{
Brendan Lu
}
\keyword{internal}
