% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpca.R
\name{.unravel_index}
\alias{.unravel_index}
\title{R implementation of np.unravel_index. NOTE: currently only works for 1D to 2D
column-major conversion, and returns a list of 2D indices. Returns a matrix
output of length(indices) columns, with two rows. The first row corresponds
to the sorted k indices, and the second row contains the sorted t indices.}
\usage{
.unravel_index(indices, dim)
}
\description{
R implementation of np.unravel_index. NOTE: currently only works for 1D to 2D
column-major conversion, and returns a list of 2D indices. Returns a matrix
output of length(indices) columns, with two rows. The first row corresponds
to the sorted k indices, and the second row contains the sorted t indices.
}
\author{
Brendan Lu
}
\keyword{internal}
