% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpls.R
\name{.validate_tpls_x_y_dim}
\alias{.validate_tpls_x_y_dim}
\title{Run some sense checks on x and y tensor inputs for tpls}
\usage{
.validate_tpls_x_y_dim(x, y)
}
\description{
Run some sense checks on x and y tensor inputs for tpls
}
\author{
Brendan Lu
}
\keyword{internal}
