% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tplsda.R
\name{.y_to_tens}
\alias{.y_to_tens}
\title{Check if y input is a matrix or tensor that can be inferred as repeated class
measurements for each sample; if so, return these classes in a consistent
form as a n x 1 x t tensor.}
\usage{
.y_to_tens(y)
}
\description{
Note: We do not have to check that the dimensions are compatible with x
tensor input, this check will be done in the call to tpls.
}
\author{
Brendan Lu
}
\keyword{internal}
