% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.mproduct.R
\name{facewise_crossproduct}
\alias{facewise_crossproduct}
\title{Tensor cross product}
\usage{
facewise_crossproduct(a, b)
}
\arguments{
\item{a}{Tensor input.}

\item{b}{Tensor input.}
}
\value{
Tensor facewise cross product.
}
\description{
Compute the equivalent of ft(a) %fp% b
}
\author{
Brendan Lu
}
