% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.mproduct.R
\name{facewise_product}
\alias{facewise_product}
\alias{\%fp\%}
\title{Tensor facewise product}
\usage{
facewise_product(..., bpparam = NULL)

a \%fp\% b
}
\arguments{
\item{...}{Arbitrary number of numerical tensor inputs.}

\item{bpparam}{A \linkS4class{BiocParallelParam} object indicating the type
of parallelisation.}
}
\value{
Cumulative facewise product.
}
\description{
Compute Kilmer's facewise product cumulatively across any arbitrary number of
tensor inputs.
}
\section{Functions}{
\itemize{
\item \code{a \%fp\% b}: Custom facewise product operator

}}
\author{
Brendan Lu
}
