% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.mproduct.R
\name{facewise_transpose}
\alias{facewise_transpose}
\alias{ft}
\title{Facewise transpose an order-3 tensor}
\usage{
facewise_transpose(tensor)

ft(tensor)
}
\arguments{
\item{tensor}{Numerical 3D array input.}
}
\value{
Facewise transpose of \code{tensor}
}
\description{
Facewise transpose an order-3 tensor
}
\section{Functions}{
\itemize{
\item \code{ft()}: Alias for \code{\link{facewise_product}}

}}
\author{
Brendan Lu
}
