% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.nipals.R
\name{impute.nipals}
\alias{impute.nipals}
\title{Impute missing values using NIPALS algorithm}
\usage{
impute.nipals(X, ncomp, ...)
}
\arguments{
\item{X}{A numeric matrix containing missing values}

\item{ncomp}{Positive integer}

\item{...}{Optional arguments passed to \code{\link[mixOmics]{nipals}}}
}
\value{
A numeric matrix with missing values imputed.
}
\description{
Impute missing values using NIPALS algorithm
}
\examples{
data("nutrimouse")
X <- data.matrix(nutrimouse$lipid)
## add missing values to X to impute and compare to actual values
set.seed(42)
na.ind <- sample(seq_along(X), size = 10)
true.values <- X[na.ind]
X[na.ind] <- NA
X.impute <- impute.nipals(X = X, ncomp = 5)
## compare
round(X.impute[na.ind], 2)
true.values
}
\author{
Al J Abadi
}
