\name{internal-functions}
\encoding{latin1}
\alias{loo}
\alias{Mfold}
\alias{bin.color}
\alias{map}
\alias{pcasvd}
\alias{ica.def}
\alias{ica.par}
% from the helpers file:
\alias{soft_thresholding_L1}
\alias{soft.threshold}
\alias{sparsity}
\alias{l2.norm}
% borrowed from RGCCA
\alias{cov2}
\alias{scale2}
\alias{initsvd}
\alias{miscrossprod}
\alias{defl.select}
\alias{deflation}


\title{Internal Functions}

\description{
Internal functions not to be used by the user.
}
