% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_test.keepX}
\alias{.check_test.keepX}
\alias{.check_ncomp}
\title{Check test.keepX}
\usage{
.check_test.keepX(test.keepX, X, indY = NULL, already.tested.X = NULL)

.check_ncomp(ncomp, X, default = 2)
}
\arguments{
\item{test.keepX}{test.keepX}

\item{X}{A matrix or a list of matrices with \code{dim} method}

\item{indY}{indY}

\item{already.tested.X}{already.tested.X}

\item{ncomp}{ncomp arg}

\item{default}{Integer, default value if \code{ncomp} is \code{NULL}}
}
\value{
test.keepX, possibly re-ordered by names for list X

Integer, or a condition
}
\description{
Check test.keepX is valid when X is either matrix or list

Check that ncomp is positive integer <= smallest dim in the data
}
\examples{

\dontrun{
.check_ncomp(300, X = mtcars)
#> Error in .check_ncomp(300, X = mtcars) : 
#>     'ncomp' must be smaller than or equal to the smallest dimenion in X: 11
.check_ncomp(NULL, X = mtcars, default = 3)
#> 3
}

}
\keyword{Internal}
