\name{linnerud}
\encoding{UTF-8}
\alias{linnerud}

\docType{data}

\title{Linnerud Dataset}

\description{
Three physiological and three exercise variables are measured on twenty middle-aged
men in a fitness club.
}

\usage{data(linnerud)}

\format{
A list containing the following components:
\describe{
\item{\code{exercise}}{data frame with 20 observations on 3 exercise variables.}
\item{\code{physiological}}{data frame with 20 observations on 3 physiological variables.}
}
}

\source{
Tenenhaus, M. (1998), Table 1, page 15.
}

\references{
Tenenhaus, M. (1998). \emph{La regression PLS: theorie et pratique}. Paris: Editions Technic.
}

\keyword{datasets}
\value{none}
