% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.mproduct.R
\name{m_product}
\alias{m_product}
\title{Kilmer's tensor-tensor m-product}
\usage{
m_product(..., m = NULL, minv = NULL, bpparam = NULL)
}
\arguments{
\item{...}{Arbitrary number of numerical tensor inputs.}

\item{m}{A function which applies an orthogonal tensor tubal transform.}

\item{minv}{The inverse of m.}

\item{bpparam}{A \linkS4class{BiocParallelParam} object indicating the type
of parallelisation. Does not have any effect if transform functions
explicitly set using \code{m}, \code{minv}.}
}
\value{
Cumulative m-product.
}
\description{
This function works cumulatively across any arbitrary number of tensor
inputs.
}
\author{
Brendan Lu
}
