% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-MultiModal.R
\name{make_unique_feature_names}
\alias{make_unique_feature_names}
\title{Make unique feature names for assays}
\usage{
make_unique_feature_names(data)
}
\arguments{
\item{data}{A \code{MultiAssayExperiment} object}
}
\value{
A \code{MultiAssayExperiment} object with unique feature names
 in all assays
}
\description{
Make unique feature names for assays
}
\examples{
\dontrun{
library(MultiAssayExperiment)
data("miniACC")
miniACC <- MatchedAssayExperiment(miniACC)
make_unique_feature_names(miniACC)
}
}
