% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.mproduct.R
\name{matrix_to_m_transforms}
\alias{matrix_to_m_transforms}
\title{Create m-transform functions as defined by a matrix}
\usage{
matrix_to_m_transforms(m_mat, m_inv_mat = NULL, bpparam = NULL)
}
\arguments{
\item{m_inv_mat}{Function which defines inverse tubal transform}

\item{bpparam}{A \linkS4class{BiocParallelParam} object indicating the type
of parallelisation.}

\item{mat}{Function which defines the tubal transform.}
}
\value{
\item{m}{A function which applies the matrix m_mat along the last dimension
of a given numerical input array. For 3D tensors it performs the mode-3
product.}
\item{minv}{The inverse of m.}
}
\description{
Returns functions \code{m} and \code{m_inv} which apply tubal transforms
defined by the matrix \code{m_mat}.
}
\author{
Brendan Lu
}
