% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalSparsePLS.R
\name{MultiModalSparsePLS}
\alias{MultiModalSparsePLS}
\title{Multi-modal integration of MultiAssayExperiment assays}
\usage{
MultiModalSparsePLS(
  data,
  formula,
  keep_features = NULL,
  ncomp = 2,
  design = "full",
  scale = TRUE,
  ...
)
}
\arguments{
\item{data}{A \code{MultiAssayExperiment} object with matching cells/samples
across all experiments.}

\item{formula}{A formula of form LHS ~ RHS where LHS is an assay and RHS is 
a list of assays (e.g. assay_1 + assay_2 + ...)}

\item{keep_features}{Named list of number of features to keep on each component
and assay. See examples.}

\item{ncomp}{the number of components to include in the model. Default to 2.
Applies to all blocks.}

\item{design}{Matrix or character. Connects assays during integration.
Each value is a continuous number b/w 0 (not connected) and 1 (fully connected).
'null' will connect the LHS to all RHS only and 'full' will fully connect all
assays.}

\item{scale}{Logical. If scale = TRUE, each block is standardized to zero
means and unit variances (default: TRUE)}

\item{...}{Other arguments passed to mixOmics::block.spls}
}
\value{
A 'multimodal' object
}
\description{
Multi-modal integration of MultiAssayExperiment assays
}
\examples{
\dontrun{
library(MultiAssayExperiment)
data("miniACC")
miniACC <- MatchedAssayExperiment(miniACC)
ncomp <- 2
keep_features <- list(RNASeq2GeneNorm = rep(20, ncomp), 
                      gistict= rep(10, ncomp), 
                      miRNASeqGene= rep(30, ncomp))
keep_features <- NULL
res <- MultiModalSparsePLS(data = miniACC, 
                           formula = RNASeq2GeneNorm ~ gistict + miRNASeqGene,
                           keep_features = keep_features, ncomp = ncomp,
                           design = 'null', scale = TRUE)

## sample plot
plotIndiv(res, pch = 16, group = miniACC$pathologic_stage, legend = TRUE)
}
}
