% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalSparsePLSDA.R
\name{MultiModalSparsePLSDA}
\alias{MultiModalSparsePLSDA}
\title{Multi-modal supervised integration of MultiAssayExperiment assays}
\usage{
MultiModalSparsePLSDA(
  data,
  formula,
  keep_features = NULL,
  ncomp = 2,
  design = "full",
  scale = TRUE,
  ...
)
}
\arguments{
\item{data}{A \code{MultiAssayExperiment} object with matching cells/samples
across all experiments.}

\item{formula}{A formula of form LHS ~ RHS where LHS is a name from colData(data) 
and RHS is a list of assays (e.g. assay_1 + assay_2 + ...).}

\item{keep_features}{Named list of number of features to keep on each component
and assay. See examples.}

\item{ncomp}{the number of components to include in the model. Default to 2.
Applies to all blocks.}

\item{design}{Matrix or character. Connects assays during integration.
Each value is a continuous number b/w 0 (not connected) and 1 (fully connected).
'null' will connect the LHS to all RHS only and 'full' will fully connect all
assays.}

\item{scale}{Logical. If scale = TRUE, each block is standardized to zero
means and unit variances (default: TRUE)}

\item{...}{Other arguments passed to mixOmics::block.splsda}
}
\value{

}
\description{
Multi-modal supervised integration of MultiAssayExperiment assays
}
\examples{
\dontrun{
library(MultiAssayExperiment)
data("miniACC")
miniACC <- MatchedAssayExperiment(miniACC)
ncomp <- 2
## mixOmics needs unique feature names across modalities
miniACC <- make_unique_feature_names(miniACC)
keep_features <- list(RNASeq2GeneNorm = rep(20, ncomp), 
                      gistict= rep(10, ncomp), 
                      miRNASeqGene= rep(30, ncomp))

res <- MultiModalSparsePLSDA(data = miniACC, 
                             formula = vital_status ~
                                 RNASeq2GeneNorm + gistict + miRNASeqGene,
                             keep_features = keep_features, ncomp = ncomp,
                             design = 'full', scale = TRUE)

## sample plot
plotIndiv(res, pch = 16, group = miniACC$pathologic_stage, legend = TRUE)
circosPlot(res, cutoff = 0.8, comp = 1, block.labels.adj = -0.1, showIntraLinks = TRUE)
}
}
