% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.pca}
\alias{plot.pca}
\title{Show PCA output}
\usage{
\method{plot}{pca}(
  x,
  ncomp = min(10, length(x$sdev)),
  type = "barplot",
  explained.var = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{pca} object}

\item{ncomp}{Integer, the number of components}

\item{type}{Character, default "barplot" or any other type available in plot, as "l","b","p",..}

\item{explained.var}{Logical, whether to plot the explained variance}

\item{...}{Not used}
}
\description{
Show PCA output
}
\author{
Kim-Anh Lê Cao, Florian Rohart, Leigh Coonan, Al J Abadi
}
