% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.spca}
\alias{plot.spca}
\title{Show sPCA output}
\usage{
\method{plot}{spca}(x, ncomp = length(x$explained_variance), type = "barplot", ...)
}
\arguments{
\item{x}{A \code{spca} object}

\item{ncomp}{Integer, the number of components}

\item{type}{Character, default "barplot" or any other type available in plot, as "l","b","p",..}

\item{...}{Not used}
}
\description{
Show sPCA output
}
\author{
Al J Abadi, Kim-Anh Lê Cao, Florian Rohart, Leigh Coonan
}
