% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndiv.pls.R
\name{plotIndiv.mixo_pls}
\alias{plotIndiv.mixo_pls}
\title{PLS sample plot methods}
\usage{
\method{plotIndiv}{mixo_pls}(object, comp = NULL, rep.space = NULL,
  ind.names = TRUE, group, col.per.group, style = "ggplot2",
  ellipse = FALSE, ellipse.level = 0.95, centroid = FALSE,
  star = FALSE, title = NULL, subtitle, legend = FALSE,
  X.label = NULL, Y.label = NULL, Z.label = NULL, abline = FALSE,
  xlim = NULL, ylim = NULL, col, cex, pch, pch.levels, alpha = 0.2,
  axes.box = "box", layout = NULL, size.title = rel(2),
  size.subtitle = rel(1.5), size.xlabel = rel(1),
  size.ylabel = rel(1), size.axis = rel(0.8), size.legend = rel(1),
  size.legend.title = rel(1.1), legend.title = "Legend",
  legend.title.pch = "Legend", legend.position = "right",
  point.lwd = 1, background = NULL, ...)
}
\description{
PLS sample plot methods
}
