% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarkers.R
\name{plotMarkers}
\alias{plotMarkers}
\title{Plot the values for multivariate markers in block analyses}
\usage{
plotMarkers(
  object,
  block,
  comp = 1,
  group = NULL,
  top_n = NULL,
  col.per.group = NULL,
  contrib = c("max", "min"),
  global = FALSE,
  title = NULL
)
}
\arguments{
\item{object}{An object of class \code{block.splsda} or \code{block.spls}}

\item{block}{Name or index of the block to use}

\item{comp}{Integer, the component to use}

\item{group}{Factor, the grouping variable (only required for
\code{block.spls} objects)}

\item{top_n}{Integer, only include this number of top features}

\item{col.per.group}{character (or symbol) color to be used when 'group' is
defined. Vector of the same length as the number of groups.}

\item{contrib}{The contribution type. one of c('max', 'min')}

\item{global}{Logical indicating whether to show the global plots (TRUE) or
segregate by feature (FALSE)}

\item{title}{The plot title}
}
\value{
A ggplot object
}
\description{
Plots the standardised values (after centring and/or scaling) for the
selected variables for a given block on a given component. Only applies to
\code{block.splsda} or \code{block.spls}.
}
\examples{
# see ?block.splsda and ?block.spls
}
\seealso{
\code{\link{plotLoadings}}, \code{\link{block.splsda}}, \code{\link{block.spls}}
}
