% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSig.R
\name{plotSig}
\alias{plotSig}
\title{Plot the values for signature variables}
\usage{
plotSig(
  object,
  block,
  comp = 1,
  group = NULL,
  top_n = NULL,
  col.per.group = NULL,
  contrib = c("max", "min"),
  global = FALSE,
  title = NULL
)
}
\arguments{
\item{object}{An object of class \code{block.splsda} or \code{block.spls}}

\item{block}{Name of the block to use}

\item{comp}{Integer, the component to use}

\item{group}{Factor, the grouping variable (not required for DA objects)}

\item{top_n}{Integer, only show this number of top features}

\item{col.per.group}{character (or symbol) color to be used when 'group' is
defined. Vector of the same length as the number of groups.}

\item{contrib}{The contribution type. one of c('max', 'min')}

\item{global}{Logical indicating whether to seggregate all features or show
the global plots.}
}
\value{
A ggplot object
}
\description{
Plots the standardaised 'expression' values for the variables for
a given block on a given component.
}
\examples{
see ?block.splsda and ?block.spls
}
