% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpca.R
\name{tpca}
\alias{tpca}
\title{Tensor PCA-like dimensionality reduction}
\usage{
tpca(
  x,
  ncomp = NULL,
  m = NULL,
  minv = NULL,
  center = TRUE,
  matrix_output = TRUE,
  bpparam = NULL
)
}
\description{
Tensor analogue of PCA introduced by Mor et al. (2022) based on Kilmer's
m-product algebra and tsvdm.
}
\author{
Brendan Lu
}
