% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpls.R
\name{tpls}
\alias{tpls}
\title{Tensor PLS-like regression}
\usage{
tpls(
  x,
  y,
  ncomp = NULL,
  m = NULL,
  minv = NULL,
  mode = "regression",
  center = TRUE,
  matrix_output = TRUE,
  bpparam = NULL
)
}
\arguments{
\item{x}{Tensor input.}

\item{y}{Tensor input.}

\item{ncomp}{The estimated number of components. ncomp must be explicitly set
as an integer in tpls.}

\item{m}{A function which applies an orthogonal tensor tubal transform.}

\item{minv}{The inverse of m.}

\item{mode}{Currently supports tensor analogues of canonical, regression,
and svd PLS modes. Defaults to "regression" mode.}

\item{center}{If set to false, the data tensor will not be centralized into
Mean Deviation Form (see Mor et al. 2022). By default, the mean horizontal
slice of the input tensor(s) are subtracted, so that all of the horizontal
slices sum to 0, analgous to centering matrix data.}

\item{matrix_output}{Note: ONLY AFFECTS THE OUTPUT IN "tsvdm" MODE.
Collects the top singular vectors across the tensor, organized by the
magnitude of the corresponding singular vector, and place them into the
columns of a matrix. Corresponds to the 'matrix compression' type of scheme
described in Mor et al. 2022.}

\item{bpparam}{A \linkS4class{BiocParallelParam} object indicating the type
of parallelisation. Does not have any effect if transform functions
explicitly set using \code{m}, \code{minv}.}
}
\description{
Developed @ Melbourne Integrative Genomics
}
\author{
Brendan Lu
}
