% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tplsda.R
\name{tplsda}
\alias{tplsda}
\title{Run tensor PLSDA-like analysis. Note this always returns a compressed-matrix
form output.}
\usage{
tplsda(
  x,
  y,
  multilevel = FALSE,
  ncomp = NULL,
  m = NULL,
  minv = NULL,
  center = TRUE,
  bpparam = NULL
)
}
\arguments{
\item{x}{Tensor input.}

\item{y}{A vector / column matrix / column tensor with n class labels, or a
n x t matrix / n x 1 x t tensor if multilevel = TRUE.}

\item{multilevel}{Set to TRUE if y contains repeated class measurements
across the t timepoints specified in the x tensor.}

\item{ncomp}{The estimated number of components. ncomp must be explicitly set
as an integer in tpls.}

\item{m}{A function which applies an orthogonal tensor tubal transform.}

\item{minv}{The inverse of m.}

\item{center}{If set to false, the data tensor will not be centralized into
Mean Deviation Form (see Mor et al. 2022). By default, the mean horizontal
slice of the input tensor(s) are subtracted, so that all of the horizontal
slices sum to 0, analgous to centering matrix data.}

\item{bpparam}{A \linkS4class{BiocParallelParam} object indicating the type
of parallelisation. Does not have any effect if transform functions
explicitly set using \code{m}, \code{minv}.}
}
\description{
Developed @ Melbourne Integrative Genomics
}
\author{
Brendan Lu
}
