% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tsvdm.R
\name{tsvdm}
\alias{tsvdm}
\title{Tensor SVD-like decomposition algorithm}
\usage{
tsvdm(
  x,
  m = NULL,
  minv = NULL,
  transform = TRUE,
  keep_hats = FALSE,
  svals_matrix_form = FALSE,
  bpparam = NULL
)
}
\description{
Return the t-SVDM decomposition from Kilmer et al. (2021).
}
\author{
Brendan Lu
}
