\name{valid}
\encoding{latin1}
\alias{valid}


\title{This function has been supersided}

\description{
The valid function has been supersided by the perf function to evaluate the performance of the PLS, sPLS, PLS-DA and sPLS-DA methods. See ?perf for more details.
}

\usage{
valid(object, \ldots)	
}

\arguments{
  \item{object}{object of class inheriting from \code{"pls"}, \code{"plsda"}, 
    \code{"spls"} or \code{"splsda"}.}
  \item{...}{arguments to pass to \code{\link{nearZeroVar}}.}
}

\author{S\'ebastien D\'ejean, Ignacio Gonz\`alez and Kim-Anh L\^e Cao.}


\keyword{regression}
\keyword{multivariate}
