\name{which.function.should.I.use}
\encoding{latin1}
\alias{which.function.should.I.use}

\title{Calculation of explained variance}

\description{
This function directs you towards the functions that fit your data.
}

\usage{
which.function.should.I.use()
}	

\details{

\code{which.function.should.I.use} asks you a few questions to evaluate your needs and to direct you to the best functions of \code{mixOmics} that suits your needs.}


\author{Florian Rohart}


\examples{
\dontrun{
which.function.should.I.use()
}
}