% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-defunct.R
\name{estim.regul}
\alias{estim.regul}
\alias{image.estim.regul}
\alias{pcatune}
\title{Estimate the parameters of regularization for Regularized CCA}
\value{
none

none

none
}
\description{
This function has been renamed \code{tune.rcc}, see \code{\link{tune.rcc}}.

This function has been renamed 'image.tune.rcc', see
\code{\link{image.tune.rcc}}.

This function has been renamed \code{\link{tune.pca}}.
}
