% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.plsda.R
\docType{methods}
\name{block.plsda}
\alias{block.plsda}
\alias{block.plsda,ANY,ANY,ANY,ANY-method}
\alias{block.plsda,MultiAssayExperiment,ANY,ANY,ANY-method}
\alias{block.plsda,ANY,ANY,ANY,formula-method}
\title{N-integration with Projection to Latent Structures models (PLS) with
Discriminant Analysis}
\usage{
block.plsda(data = NULL, X = NULL, Y = NULL, formula = NULL, ...)

\S4method{block.plsda}{ANY,ANY,ANY,ANY}(data = NULL, X = NULL,
  Y = NULL, formula = NULL, ...)

\S4method{block.plsda}{MultiAssayExperiment,ANY,ANY,ANY}(data = NULL,
  X = NULL, Y = NULL, formula = NULL, ...)

\S4method{block.plsda}{ANY,ANY,ANY,formula}(data = NULL, X = NULL,
  Y = NULL, formula = NULL, ...)
}
\arguments{
\item{data}{An object of class \code{MultiAssayExperiment} or similar.
See Details.}

\item{X}{A list of data sets (called 'blocks') measured on the same samples.
Data in the list should be arranged in matrices, samples x variables, with
samples order matching in all data sets.}

\item{Y}{A factor or a class vector indicating the discrete outcome of each
sample.}

\item{formula}{A \code{formula} object of form \code{LHS ~ RHS} where 
\code{LHS} and \code{RHS} are either matrices or assay names (no quotation)
 from \code{data}. 
 \code{X} and \code{Y} must be \code{NULL} when \code{formula} is used. 
 See Details.}

\item{...}{Aguments passed to the generic.}

\item{indY}{To be supplied if Y is missing, indicates the position of the
factor / class vector outcome in the list \code{X}}
}
\value{
\code{block.plsda} returns an object of class
\code{"block.plsda","block.pls"}, a list that contains the following
components:

\item{X}{the centered and standardized original predictor matrix.}
\item{indY}{the position of the outcome Y in the output list X.}
\item{ncomp}{the number of components included in the model for each block.}
\item{mode}{the algorithm used to fit the model.} \item{variates}{list
containing the variates of each block of X.} \item{loadings}{list containing
the estimated loadings for the variates.} \item{names}{list containing the
names to be used for individuals and variables.} \item{nzv}{list containing
the zero- or near-zero predictors information.} \item{iter}{Number of
iterations of the algorthm for each component}
\item{explained_variance}{Percentage of explained variance for each
component and each block}
}
\description{
Integration of multiple data sets measured on the same samples or
observations to classify a discrete outcome, ie. N-integration with
Discriminant Analysis. The method is partly based on Generalised Canonical
Correlation Analysis.
}
\details{
\code{block.plsda} function fits a horizontal integration PLS-DA model with
a specified number of components per block). A factor indicating the
discrete outcome needs to be provided, either by \code{Y} or by its position
\code{indY} in the list of blocks \code{X}.

\code{X} can contain missing values. Missing values are handled by being
disregarded during the cross product computations in the algorithm
\code{block.pls} without having to delete rows with missing data.
Alternatively, missing data can be imputed prior using the \code{nipals}
function.

The type of algorithm to use is specified with the \code{mode} argument.
Four PLS algorithms are available: PLS regression \code{("regression")}, PLS
canonical analysis \code{("canonical")}, redundancy analysis
\code{("invariant")} and the classical PLS algorithm \code{("classic")} (see
References and \code{?pls} for more details).

Note that our method is partly based on Generalised Canonical Correlation
Analysis and differs from the MB-PLS approaches proposed by Kowalski et al.,
1989, J Chemom 3(1) and Westerhuis et al., 1998, J Chemom, 12(5).
}
\examples{
#' \dontrun{
data = list(gene = nutrimouse$gene, lipid = nutrimouse$lipid, Y = nutrimouse$diet)
# with this design, all blocks are connected
design = matrix(c(0,1,1,1,0,1,1,1,0), ncol = 3, nrow = 3,
byrow = TRUE, dimnames = list(names(data), names(data)))
## when index of Y in list of X is provided
res = block.plsda(X = data, indY = 3) # indY indicates where the outcome Y is in the list X
plotIndiv(res, ind.names = FALSE, legend = TRUE)
plotVar(res)


# when Y is provided
res = block.plsda(X = list(gene = nutrimouse$gene, lipid = nutrimouse$lipid),
Y = nutrimouse$diet, ncomp = 2)
plotIndiv(res)
plotVar(res)
#' }
}
\references{
On PLSDA:

Barker M and Rayens W (2003). Partial least squares for discrimination.
\emph{Journal of Chemometrics} \bold{17}(3), 166-173. Perez-Enciso, M. and
Tenenhaus, M. (2003). Prediction of clinical outcome with microarray data: a
partial least squares discriminant analysis (PLS-DA) approach. \emph{Human
Genetics} \bold{112}, 581-592. Nguyen, D. V. and Rocke, D. M. (2002). Tumor
classification by partial least squares using microarray gene expression
data. \emph{Bioinformatics} \bold{18}, 39-50.

On multiple integration with PLS-DA: Gunther O., Shin H., Ng R. T. ,
McMaster W. R., McManus B. M. , Keown P. A. , Tebbutt S.J. , Lê Cao K-A. ,
(2014) Novel multivariate methods for integration of genomics and proteomics
data: Applications in a kidney transplant rejection study, OMICS: A journal
of integrative biology, 18(11), 682-95.

On multiple integration with sPLS-DA and 4 data blocks:

Singh A., Gautier B., Shannon C., Vacher M., Rohart F., Tebbutt S. and Lê
Cao K.A. (2016). DIABLO: multi omics integration for biomarker discovery.
BioRxiv available here:
\url{http://biorxiv.org/content/early/2016/08/03/067611}

mixOmics article:

Rohart F, Gautier B, Singh A, Lê Cao K-A. mixOmics: an R package for 'omics
feature selection and multiple data integration. PLoS Comput Biol 13(11):
e1005752
}
\seealso{
\code{\link{plotIndiv}}, \code{\link{plotArrow}},
\code{\link{plotLoadings}}, \code{\link{plotVar}}, \code{\link{predict}},
\code{\link{perf}}, \code{\link{selectVar}}, \code{\link{block.pls}},
\code{\link{block.splsda}} and http://www.mixOmics.org for more details.
}
\author{
Florian Rohart, Benoit Gautier, Kim-Anh Lê Cao, Al J Abadi.
}
\keyword{multivariate}
\keyword{regression}
