% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.spls.R
\docType{methods}
\name{block.spls}
\alias{block.spls}
\alias{block.spls,ANY,ANY,ANY,ANY-method}
\alias{block.spls,MultiAssayExperiment,ANY,ANY,ANY-method}
\alias{block.spls,MatchedAssayExperiment,ANY,ANY,ANY-method}
\alias{block.spls,ANY,ANY,ANY,formula-method}
\title{N-integration and feature selection with sparse Projection to Latent
Structures models (sPLS)}
\usage{
block.spls(data = NULL, X = NULL, Y = NULL, formula = NULL, ...)

\S4method{block.spls}{ANY,ANY,ANY,ANY}(data = NULL, X = NULL,
  Y = NULL, formula = NULL, ...)

\S4method{block.spls}{MultiAssayExperiment,ANY,ANY,ANY}(data = NULL,
  X = NULL, Y = NULL, formula = NULL, ...)

\S4method{block.spls}{MatchedAssayExperiment,ANY,ANY,ANY}(data = NULL,
  X = NULL, Y = NULL, formula = NULL, ...)

\S4method{block.spls}{ANY,ANY,ANY,formula}(data = NULL, X = NULL,
  Y = NULL, formula = NULL, ...)
}
\arguments{
\item{keepX}{A list of same length as X.  Each entry is the number of
variables to select in each of the blocks of X for each component. By
default all variables are kept in the model.}

\item{keepY}{Only if Y is provided. Each entry is the number of variables to
select in each of the blocks of Y for each component.}
}
\value{
\code{block.spls} returns an object of class \code{"block.spls"}, a
list that contains the following components:

\item{X}{the centered and standardized original predictor matrix.}
\item{indY}{the position of the outcome Y in the output list X.}
\item{ncomp}{the number of components included in the model for each block.}
\item{mode}{the algorithm used to fit the model.} \item{keepX}{Number of
variables used to build each component of each block} \item{keepY}{Number of
variables used to build each component of Y} \item{variates}{list containing
the variates of each block of X.} \item{loadings}{list containing the
estimated loadings for the variates.} \item{names}{list containing the names
to be used for individuals and variables.} \item{nzv}{list containing the
zero- or near-zero predictors information.} \item{iter}{Number of iterations
of the algorthm for each component} \item{explained_variance}{Percentage of
explained variance for each component and each block}
}
\description{
Integration of multiple data sets measured on the same samples or
observations, with variable selection in each data set, ie. N-integration.
The method is partly based on Generalised Canonical Correlation Analysis.
}
\details{
\code{block.spls} function fits a horizontal sPLS model with a specified
number of components per block). An outcome needs to be provided, either by
\code{Y} or by its position \code{indY} in the list of blocks \code{X}.
Multi (continuous)response are supported. \code{X} and \code{Y} can contain
missing values. Missing values are handled by being disregarded during the
cross product computations in the algorithm \code{block.pls} without having
to delete rows with missing data. Alternatively, missing data can be imputed
prior using the \code{nipals} function.

The type of algorithm to use is specified with the \code{mode} argument.
Four PLS algorithms are available: PLS regression \code{("regression")}, PLS
canonical analysis \code{("canonical")}, redundancy analysis
\code{("invariant")} and the classical PLS algorithm \code{("classic")} (see
References and \code{?pls} for more details).

Note that our method is partly based on sparse Generalised Canonical
Correlation Analysis and differs from the MB-PLS approaches proposed by
Kowalski et al., 1989, J Chemom 3(1), Westerhuis et al., 1998, J Chemom,
12(5) and sparse variants Li et al., 2012, Bioinformatics 28(19); Karaman et
al (2014), Metabolomics, 11(2); Kawaguchi et al., 2017, Biostatistics.

Variable selection is performed on each component for each block of
\code{X}, and for \code{Y} if specified, via input parameter \code{keepX}
and \code{keepY}.

Note that if \code{Y} is missing and \code{indY} is provided, then variable
selection on \code{Y} is performed by specifying the input parameter
directly in \code{keepX} (no \code{keepY} is needed).
}
\examples{
# Example with multi omics TCGA study
# -----------------------------
# this is the X data as a list of mRNA and miRNA; the Y data set is a single data set of proteins
data = list(mrna = breast.TCGA$data.train$mrna, mirna = breast.TCGA$data.train$mirna)
# set up a full design where every block is connected
design = matrix(1, ncol = length(data), nrow = length(data),
dimnames = list(names(data), names(data)))
diag(design) =  0
design
# set number of component per data set
ncomp = c(2)
# set number of variables to select, per component and per data set (this is set arbitrarily)
list.keepX = list(mrna = rep(20, 2), mirna = rep(10,2))
list.keepY = c(rep(10, 2))

TCGA.block.spls = block.spls(X = data, Y = breast.TCGA$data.train$protein,
ncomp = ncomp, keepX = list.keepX, keepY = list.keepY, design = design)
TCGA.block.spls
# in plotindiv we color the samples per breast subtype group but the method is unsupervised!
plotIndiv(TCGA.block.spls, group =  breast.TCGA$data.train$subtype, ind.names = FALSE)
# illustrates coefficient weights in each block
plotLoadings(TCGA.block.spls, ncomp = 1)
plotVar(TCGA.block.spls, style = 'graphics', legend = TRUE)
network(TCGA.block.spls)

# Example with MultiAssayExperiments
# -----------------------------------
library(mixOmicsData)
data("breast.TCGA.mae")

TCGA.block.spls.mae <- block.spls(data = breast.TCGA.mae, keepX = list.keepX, keepY = list.keepY,
                                formula = protein ~ mrna + mirna, ncomp = ncomp,
                                design = design)

## TODO add more examples
}
\references{
Tenenhaus, M. (1998). \emph{La regression PLS: theorie et
pratique}. Paris: Editions Technic.

Wold H. (1966). Estimation of principal components and related models by
iterative least squares. In: Krishnaiah, P. R. (editors), \emph{Multivariate
Analysis}. Academic Press, N.Y., 391-420.

Tenenhaus A. and Tenenhaus M., (2011), Regularized Generalized Canonical
Correlation Analysis, Psychometrika, Vol. 76, Nr 2, pp 257-284.

Tenenhaus A., Philippe C., Guillemot V, Lê Cao K.A., Grill J, Frouin V.
Variable selection for generalized canonical correlation analysis.
\emph{Biostatistics}. kxu001
}
\seealso{
\code{\link{plotIndiv}}, \code{\link{plotArrow}},
\code{\link{plotLoadings}}, \code{\link{plotVar}}, \code{\link{predict}},
\code{\link{perf}}, \code{\link{selectVar}}, \code{\link{block.pls}},
\code{\link{block.splsda}} and http://www.mixOmics.org for more details.
}
\author{
Florian Rohart, Benoit Gautier, Kim-Anh Lê Cao, Al J Abadi.
}
\keyword{multivariate}
\keyword{regression}
