% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circosPlot.R
\name{circosPlot}
\alias{circosPlot}
\title{circosPlot for DIABLO}
\usage{
circosPlot(object, comp = 1:min(object$ncomp), cutoff, color.Y,
  color.blocks, color.cor, var.names = NULL, showIntraLinks = FALSE,
  line = TRUE, size.legend = 0.8, ncol.legend = 1,
  size.variables = 0.25, size.labels = 1, legend = TRUE)
}
\arguments{
\item{object}{An object of class inheriting from \code{"block.splsda"}.}

\item{comp}{Numeric vector indicating which component to plot. Default to
all}

\item{cutoff}{Only shows links with a correlation higher than \code{cutoff}}

\item{color.Y}{a character vector of colors to be used for the levels of the
outcome}

\item{color.blocks}{a character vector of colors to be used for the blocks}

\item{color.cor}{a character vector of two colors. First one is for the
negative correlation, second one is for the positive correlation}

\item{var.names}{Optional parameter. A list of length the number of blocks
in \code{object$X}, containing the names of the variables of each block. If
\code{NULL}, the colnames of the data matrix are used.}

\item{showIntraLinks}{if TRUE, shows the correlation higher than the
threshold inside each block.}

\item{line}{if TRUE, shows the overall expression of the selected variables.
see examples.}

\item{size.legend}{size of the legend}

\item{ncol.legend}{number of columns for the legend}

\item{size.variables}{size of the variable labels}

\item{size.labels}{size of the block labels}

\item{legend}{boolean. Whether the legend should be added. Default is TRUE.}
}
\value{
If saved in an object, the circos plot will output the similarity
matrix and the names of the variables displayed on the plot (see
\code{attributes(object)}).
}
\description{
Displays variable correlation among different blocks
}
\details{
\code{circosPlot} function depicts correlations of variables selected with
\code{block.splsda} among different blocks, using a generalisation of the
method presented in González et al 2012. If \code{ncomp} is specified, then
only the variables selected on that component are displayed.
}
\examples{

Y = nutrimouse$diet
data = list(gene = nutrimouse$gene, lipid = nutrimouse$lipid)
design = matrix(c(0,1,1,1,0,1,1,1,0), ncol = 3, nrow = 3, byrow = TRUE)


nutrimouse.sgccda <- wrapper.sgccda(X=data,
Y = Y,
design = design,
keepX = list(gene=c(10,10), lipid=c(15,15)),
ncomp = 2,
scheme = "horst")

circosPlot(nutrimouse.sgccda, cutoff = 0.7, ncol.legend = 2, size.legend = 1.1)

\dontrun{
circosPlot(nutrimouse.sgccda, cutoff = 0.7, ncol.legend = 2, size.legend = 1.1,
color.Y = 1:5, color.blocks = c("green","brown"), color.cor = c("magenta", "purple"))

par(mfrow=c(2,2))
circosPlot(nutrimouse.sgccda, cutoff = 0.7, ncol.legend = 2,
size.legend = 1.1)
circosPlot(nutrimouse.sgccda, cutoff = 0.7, ncol.legend = 2,
size.legend = 1.1, showIntraLinks = TRUE)
circosPlot(nutrimouse.sgccda, cutoff = 0.7, ncol.legend = 1,
size.legend = 1.1, showIntraLinks = TRUE)
circosPlot(nutrimouse.sgccda, cutoff = 0.7, ncol.legend = 2,
size.legend = 1.1, showIntraLinks = TRUE, line = FALSE, size.variables = 0.5)
}

}
\references{
Singh A., Gautier B., Shannon C., Vacher M., Rohart F., Tebbutt
S. and Lê Cao K.A. (2016). DIABLO: multi omics integration for biomarker
discovery. BioRxiv available here:
\url{http://biorxiv.org/content/early/2016/08/03/067611}

mixOmics article:

Rohart F, Gautier B, Singh A, Lê Cao K-A. mixOmics: an R package for 'omics
feature selection and multiple data integration. PLoS Comput Biol 13(11):
e1005752

González I., Lê Cao K.A., Davis M.J., Déjean S. (2012). Visualising
associations between paired 'omics' data sets. \emph{BioData Mining};
\bold{5}(1).
}
\seealso{
\code{\link{block.splsda}}, references and
http://www.mixOmics.org/mixDIABLO for more details.
}
\author{
Michael Vacher, Amrit Singh, Florian Rohart, Kim-Anh Lê Cao
}
\keyword{multivariate}
\keyword{regression}
