\name{internal-functions}
\encoding{latin1}
\alias{loo}
\alias{Mfold}
\alias{bin.color}
\alias{map}
\alias{unmap}
\alias{pcasvd}



\alias{ica.def}
\alias{ica.par}

\alias{check.one.level}
\alias{Split.variation.one.level}
\alias{Split.variation.two.level}


\title{Internal Functions}

\description{
Internal functions not to be used by the user.
}
