\name{internal-functions}
\encoding{latin1}
\alias{loo}
\alias{Mfold}
\alias{bin.color}
\alias{map}
\alias{unmap}
\alias{pcasvd}

\alias{spls.model}
\alias{splsda.model}


\alias{ica.def}
\alias{ica.par}

\alias{Split.variation.one.level}
\alias{Split.variation.two.level}


\title{Internal Functions}

\description{
Internal functions not to be used by the user.
}
