\name{internal-functions}
\encoding{latin1}
\alias{loo}
\alias{Mfold}
\alias{bin.color}
\alias{map}
\alias{unmap}
\alias{pcasvd}
\alias{q2.spls}
\alias{q2.pls}

\alias{ica.def}
\alias{ica.par}



\title{Internal Functions}

\description{
Internal functions not to be used by the user.
}
