% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods-plot.R
\name{plot.pca}
\alias{plot.pca}
\title{Show (s)pca output}
\usage{
\method{plot}{pca}(x, ncomp = length(x$explained_variance), type = "barplot", ...)
}
\arguments{
\item{x}{A \code{(s)pca} object}

\item{ncomp}{Integer, the number of components}

\item{type}{Character, default "barplot" or any other type available in plot, as "l","b","p",..}

\item{...}{Not used}
}
\description{
Show (s)pca output
}
\author{
Kim-Anh Lê Cao, Florian Rohart, Leigh Coonan, Al J Abadi
}
