% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.rcc}
\alias{plot.rcc}
\title{Canonical Correlations Plot}
\usage{
\method{plot}{rcc}(x, scree.type = c("pointplot", "barplot"), ...)
}
\arguments{
\item{x}{object of class inheriting from \code{"rcc"}.}

\item{scree.type}{character string, (partially) matching one of
\code{"pointplot"} or \code{"barplot"}, determining the kind of scree plots
to be produced.}

\item{...}{arguments to be passed to other methods. For the
\code{"pointplot"} type see \code{\link{points}}, for \code{"barplot"} type
see \code{\link{barplot}}.}
}
\value{
none
}
\description{
This function provides scree plot of the canonical correlations.
}
\examples{

data(nutrimouse)
X <- nutrimouse$lipid
Y <- nutrimouse$gene
nutri.res <- rcc(X, Y, lambda1 = 0.064, lambda2 = 0.008)

## 'pointplot' type scree
plot(nutri.res) #(default)

\dontrun{
plot(nutri.res, pch = 19, cex = 1.2,
col = c(rep("red", 3), rep("darkblue", 18)))

## 'barplot' type scree
plot(nutri.res, scree.type = "barplot")

plot(nutri.res, scree.type = "barplot", density = 20, col = "black")
}

}
\seealso{
\code{\link{points}}, \code{\link{barplot}}, \code{\link{par}}.
}
\author{
Sébastien Déjean, Ignacio González, Al J Abadi
}
\keyword{hplot}
\keyword{multivariate}
