% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plsda}
\alias{plsda}
\title{Partial Least Squares Discriminant Analysis (PLS-DA).}
\usage{
\S4method{plsda}{ANY}(X, Y, ncomp = 2, scale = TRUE,
mode = c("regression", "canonical", "invariant", "classic"), tol = 1e-06,
max.iter = 100, near.zero.var = FALSE, logratio = c("none", "CLR"),
multilevel = NULL, all.outputs = TRUE)
}
\arguments{
\item{X}{A numeric matrix (or data frame) which provides the data for the
principal components analysis. It can contain missing values.
Alternatively, name of an assay from \code{data}.}

\item{Y}{a factor or a class vector for the discrete outcome.}

\item{formula}{A \code{formula} object of form \code{LHS ~ RHS} where 
\code{LHS} and \code{RHS} are either matrices or assay names (no quotation)
 from \code{data}. 
 \code{X} and \code{Y} must be \code{NULL} when \code{formula} is used. 
 See Details.}

\item{mode}{Character string. What type of algorithm to use, (partially)
matching one of \code{"regression"}, \code{"canonical"}, \code{"invariant"}
or \code{"classic"}. See Details.}

\item{near.zero.var}{Boolean, see the internal \code{\link{nearZeroVar}}
function (should be set to TRUE in particular for data with many zero
values). Setting this argument to FALSE (when appropriate) will speed up the
computations. Default value is FALSE}

\item{logratio}{One of ('none','CLR'). Default to 'none'}

\item{...}{Aguments passed to the generic.}

\item{multilevel}{sample information for multilevel decomposition for
repeated measurements. A numeric matrix or data frame indicating the
repeated measures on each individual, i.e. the individuals ID. See examples
in \code{?splsda}.}
}
\value{
\code{plsda} returns an object of class \code{"plsda"}, a list that
contains the following components:

\item{X}{the centered and standardized original predictor matrix.}
\item{Y}{the centered and standardized indicator response vector or matrix.}
\item{ind.mat}{the indicator matrix.} \item{ncomp}{the number of components
included in the model.} \item{variates}{list containing the \code{X} and
\code{Y} variates.} \item{loadings}{list containing the estimated loadings
for the variates.} \item{names}{list containing the names to be used for
individuals and variables.} \item{nzv}{list containing the zero- or
near-zero predictors information.} \item{tol}{the tolerance used in the
iterative algorithm, used for subsequent S3 methods} \item{max.iter}{the
maximum number of iterations, used for subsequent S3 methods}
\item{iter}{Number of iterations of the algorthm for each component}
\item{explained_variance}{amount of variance explained per component (note
that contrary to PCA, this amount may not decrease as the aim of the method
is not to maximise the variance, but the covariance between X and the dummy
matrix Y).} \item{mat.c}{matrix of coefficients from the regression of X /
residual matrices X on the X-variates, to be used internally by
\code{predict}.} \item{defl.matrix}{residual matrices X for each dimension.}
}
\description{
Function to perform standard Partial Least Squares regression to classify
samples.
}
\details{
\code{plsda} function fit PLS models with \eqn{1,...,}\code{ncomp}
components to the factor or class vector \code{Y}. The appropriate indicator
matrix is created.

logratio transform and multilevel analysis are performed sequentially as
internal pre-processing step, through \code{\link{logratio.transfo}} and
\code{\link{withinVariation}} respectively.

Logratio can only be applied if the data do not contain any 0 value (for
count data, we thus advise the normalise raw data with a 1 offset).

More details about the PLS modes in \code{?pls}.
}
\examples{
#' \dontrun{

data(breast.tumors)

## ---------------- with X as matrix and and Y as numeric/factor
X <- breast.tumors$gene.exp
Y <- breast.tumors$sample$treatment
plsda.res1 <- plsda(X=X, Y=Y)
plotVar(plsda.res1, cutoff = 0.6)
plotLoadings(plsda.res1)

## ---------------- formula method for matrices
## 'formula' argument should be explicitly mentioned (formula = ...)
## for correct method dispatch
plsda.res2 <- plsda(formula = Y ~ X)
## exclude calls and see if all outputs  are identical
identical(plsda.res1[-1], plsda.res2[-1])
#> TRUE
## ---------------- MultiAssayExperiment and assay names as X and Y
## 'data' argument should be explicitly mentioned for correct method dispatch
plsda.res3 <- plsda(X='exercise', Y='physiological', data = linnerud.mae)
identical(plsda.res1[-1], plsda.res3[-1])
#> TRUE

## ---------------- MultiAssayExperiment and formula with assay names
plsda.res4 <- plsda(formula = physiological ~ exercise, data = linnerud.mae, mode = "classic",)
identical(plsda.res1[-1], plsda.res4[-1])
#> TRUE

## ---------------- MultiAssayExperiment; X=assay and Y=colData
toxicity.plsda1 <- plsda(data = liver.toxicity.mae,  formula = Dose.Group~gene, ncomp = 3)
toxicity.plsda2 <- plsda(data = liver.toxicity.mae,  Y='Dose.Group', X='gene', ncomp = 3)
identical(toxicity.plsda1[-1], toxicity.plsda2[-1])
#> TRUE

#' }
}
\references{
On PLSDA: Barker M and Rayens W (2003). Partial least squares
for discrimination. \emph{Journal of Chemometrics} \bold{17}(3), 166-173.
Perez-Enciso, M. and Tenenhaus, M. (2003). Prediction of clinical outcome
with microarray data: a partial least squares discriminant analysis (PLS-DA)
approach. \emph{Human Genetics} \bold{112}, 581-592. Nguyen, D. V. and
Rocke, D. M. (2002). Tumor classification by partial least squares using
microarray gene expression data. \emph{Bioinformatics} \bold{18}, 39-50. On
log ratio transformation: Filzmoser, P., Hron, K., Reimann, C.: Principal
component analysis for compositional data with outliers. Environmetrics
20(6), 621-632 (2009) Lê Cao K.-A., Costello ME, Lakis VA, Bartolo, F,Chua
XY, Brazeilles R, Rondeau P. MixMC: Multivariate insights into Microbial
Communities. PLoS ONE, 11(8): e0160169 (2016). On multilevel decomposition:
Westerhuis, J.A., van Velzen, E.J., Hoefsloot, H.C., Smilde, A.K.:
Multivariate paired data analysis: multilevel plsda versus oplsda.
Metabolomics 6(1), 119-128 (2010) Liquet, B., Lê Cao K.-A., Hocini, H.,
Thiebaut, R.: A novel approach for biomarker selection and the integration
of repeated measures experiments from two assays. BMC bioinformatics 13(1),
325 (2012)
}
\seealso{
\code{\link{splsda}}, \code{\link{summary}},
\code{\link{plotIndiv}}, \code{\link{plotVar}}, \code{\link{predict}},
\code{\link{perf}}, \code{\link{mint.block.plsda}},
\code{\link{block.plsda}} and http://mixOmics.org for more details.
}
\author{
Ignacio González, Kim-Anh Lê Cao.
}
\keyword{multivariate}
\keyword{regression}
