% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc.R
\name{rcc}
\alias{rcc}
\alias{rcc.default}
\title{Regularized Canonical Correlation Analysis}
\usage{
rcc(X, Y, ncomp = 2, method = "ridge", lambda1 = 0, lambda2 = 0)
}
\arguments{
\item{X}{numeric matrix or data frame \eqn{(n \times p)}, the observations
on the \eqn{X} variables. \code{NA}s are allowed.}

\item{Y}{numeric matrix or data frame \eqn{(n \times q)}, the observations
on the \eqn{Y} variables. \code{NA}s are allowed.}

\item{ncomp}{the number of components to include in the model. Default to 2.}

\item{method}{One of "ridge" or "shrinkage". If "ridge", \code{lambda1} and
\code{lambda2} need to be supplied (see also our function tune.rcc); if
"shrinkage", parameters are directly estimated with Strimmer's formula, see
below and reference.}

\item{lambda1, lambda2}{a non-negative real. The regularization parameter for
the \emph{X} and \emph{Y} data. Defaults to \code{lambda1=lambda2=0}. Only
used if \code{method="ridge"}}
}
\value{
\code{rcc} returns a object of class \code{"rcc"}, a list that
contains the following components: \item{X}{the original \eqn{X} data.}
\item{Y}{the original \eqn{Y} data.} \item{cor}{a vector containing the
canonical correlations.} \item{lambda}{a vector containing the
regularization parameters whether those were input if ridge method or
directly estimated with the shrinkage method.} \item{loadings}{list
containing the estimated coefficients used to calculate the canonical
variates in \eqn{X} and \eqn{Y}.} \item{variates}{list containing the
canonical variates.} \item{names}{list containing the names to be used for
individuals and variables.}
}
\description{
The function performs the regularized extension of the Canonical Correlation
Analysis to seek correlations between two data matrices.
}
\details{
The main purpose of Canonical Correlations Analysis (CCA) is the exploration
of sample correlations between two sets of variables \eqn{X} and \eqn{Y}
observed on the same individuals (experimental units) whose roles in the
analysis are strictly symmetric.

The \code{cancor} function performs the core of computations but additional
tools are required to deal with data sets highly correlated (nearly
collinear), data sets with more variables than units by example.

The \code{rcc} function, the regularized version of CCA, is one way to deal
with this problem by including a regularization step in the computations of
CCA. Such a regularization in this context was first proposed by Vinod
(1976), then developped by Leurgans \emph{et al.} (1993). It consists in the
regularization of the empirical covariances matrices of \eqn{X} and \eqn{Y}
by adding a multiple of the matrix identity, that is, Cov\eqn{(X)+ \lambda_1
I} and Cov\eqn{(Y)+ \lambda_2 I}.

When \code{lambda1=0} and \code{lambda2=0}, \code{rcc} performs a classical
CCA, if possible (i.e. when \eqn{n > p+q}.

The shrinkage estimates \code{method = "shrinkage"} can be used to bypass
\code{\link{tune.rcc}} to choose the shrinkage parameters - which can be
long and costly to compute with very large data sets. Note that both
functions \code{\link{tune.rcc}} (which uses cross-validation) and the
whrinkage parameters (which uses the formula from Schafer and Strimmer) may
output different results.

Note: when \code{method = "shrinkage"} the input data are centered and
scaled for the estimation of the shrinkage parameters and the calculation of
the regularised variance-covariance matrices in \code{\link{rcc}}.

The estimation of the missing values can be performed by the reconstitution
of the data matrix using the \code{nipals} function. Otherwise, missing
values are handled by casewise deletion in the \code{rcc} function.
}
\examples{

## Classic CCA
X <- linnerud$exercise
Y <- linnerud$physiological
linn.res <- rcc(X, Y)

\dontrun{
## Regularized CCA
X <- nutrimouse$lipid
Y <- nutrimouse$gene
nutri.res1 <- rcc(X, Y, ncomp = 3, lambda1 = 0.064, lambda2 = 0.008)

## using shrinkage parameters
nutri.res2 <- rcc(X, Y, ncomp = 3, method = 'shrinkage')
nutri.res2$lambda # the shrinkage parameters
}

}
\references{
González, I., Déjean, S., Martin, P. G., and Baccini, A. (2008).
CCA: An R package to extend canonical correlation analysis. Journal of
Statistical Software, 23(12), 1-14.

González, I., Déjean, S., Martin, P., Goncalves, O., Besse, P., and Baccini,
A. (2009). Highlighting relationships between heterogeneous biological data
through graphical displays based on regularized canonical correlation
analysis. Journal of Biological Systems, 17(02), 173-199.

Leurgans, S. E., Moyeed, R. A. and Silverman, B. W. (1993). Canonical
correlation analysis when the data are curves. \emph{Journal of the Royal
Statistical Society. Series B} \bold{55}, 725-740.

Vinod, H. D. (1976). Canonical ridge and econometrics of joint production.
\emph{Journal of Econometrics} \bold{6}, 129-137.

Opgen-Rhein, R., and K. Strimmer. 2007. Accurate ranking of differentially
expressed genes by a distribution-free shrinkage approach. Statist.
emphAppl. Genet. Mol. Biol. \bold{6}:9.
(http://www.bepress.com/sagmb/vol6/iss1/art9/)

Sch"afer, J., and K. Strimmer. 2005. A shrinkage approach to large-scale
covariance estimation and implications for functional genomics. Statist.
emphAppl. Genet. Mol. Biol.  \bold{4}:32.
(http://www.bepress.com/sagmb/vol4/iss1/art32/)
}
\seealso{
\code{\link{summary}}, \code{\link{tune.rcc}},
\code{\link{plot.rcc}}, \code{\link{plotIndiv}}, \code{\link{plotVar}},
\code{\link{cim}}, \code{\link{network}} and http://www.mixOmics.org for
more details.
}
\author{
Sébastien Déjean, Ignacio González, Francois Bartolo.
}
\keyword{multivariate}
