% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\docType{methods}
\name{spca}
\alias{spca}
\alias{spca,MultiAssayExperiment-method}
\alias{spca,MatchedAssayExperiment-method}
\title{Sparse Principal Components Analysis}
\usage{
spca(data = NULL, X = NULL, ...)

\S4method{spca}{MultiAssayExperiment}(data = NULL, X = NULL, ...)

\S4method{spca}{MatchedAssayExperiment}(data = NULL, X = NULL, ...)
}
\arguments{
\item{data}{An object of class \code{MultiAssayExperiment} or similar.
See Details.}

\item{X}{A numeric matrix (or data frame) which provides the data for the
principal components analysis. It can contain missing values.
Alternatively, name of an assay from \code{data}.}

\item{...}{Aguments passed to the generic.}

\item{keepX}{numeric vector of length ncomp, the number of variables to keep.}

\item{logratio}{one of ('none','CLR'). Specifies the log ratio
transformation to deal with compositional values that may arise from
specific normalisation in sequencing data. Default to 'none'.
in loading vectors. By default all variables are kept in the model. See
details.}
}
\value{
\code{spca} returns a list with class \code{"spca"} containing the
following components: \item{ncomp}{the number of components to keep in the
calculation.} \item{varX}{the adjusted cumulative percentage of variances
explained.} \item{keepX}{the number of variables kept in each loading
vector.} \item{iter}{the number of iterations needed to reach convergence
for each component.} \item{rotation}{the matrix containing the sparse
loading vectors.} \item{x}{the matrix containing the principal components.}
Refer to \code{pca} for different type of inputs supported.
}
\description{
Performs a sparse principal components analysis to perform variable
selection by using singular value decomposition.

The calculation employs singular value decomposition of the (centered and
scaled) data matrix and LASSO to generate sparsity on the loading vectors.

\code{scale= TRUE} is highly recommended as it will help obtaining
orthogonal sparse loading vectors.

\code{keepX} is the number of variables to keep in loading vectors. The
difference between number of columns of \code{X} and \code{keepX} is the
degree of sparsity, which refers to the number of zeros in each loading
vector.

Note that \code{spca} does not apply to the data matrix with missing values.
The biplot function for \code{spca} is not available.

According to Filzmoser et al., a ILR log ratio transformation is more
appropriate for PCA with compositional data. Both CLR and ILR are valid.

Logratio transform and multilevel analysis are performed sequentially as
internal pre-processing step, through \code{\link{logratio.transfo}} and
\code{\link{withinVariation}} respectively.

Logratio can only be applied if the data do not contain any 0 value (for
count data, we thus advise the normalise raw data with a 1 offset). For ILR
transformation and additional offset might be needed.
}
\examples{
#' \dontrun{

data(liver.toxicity)
spca.rat <- spca(liver.toxicity$gene, ncomp = 3, keepX = rep(50, 3))
print(spca.rat)
## variable representation
plotVar(spca.rat, cex=2)
## 3d plot
plotVar(spca.rat,style="3d")
# }

# example with MultiAssayExperiment class
# --------------------------------
## X is the name of an assay from data
spca.rat <- spca(data=liver.toxicity.mae, X='gene', ncomp = 3, keepX = rep(50, 3))
plotVar(spca.rat, cex=2)
# \dontrun{


## samples representation
plotIndiv(spca.rat, ind.names = liver.toxicity$treatment[, 3],
          group = as.numeric(liver.toxicity$treatment[, 3]))
plotIndiv(spca.rat, cex = 0.01,
          col = as.numeric(liver.toxicity$treatment[, 3]),style="3d")


# example with multilevel decomposition and CLR log ratio transformation
# ----------------

data("diverse.16S")
spca.res = spca(X = diverse.16S$data.TSS, ncomp = 5, keepX = rep(20, 5),
              logratio = 'CLR', multilevel = diverse.16S$sample)
plotIndiv(spca.res, ind.names = FALSE, group = diverse.16S$bodysite, title = '16S diverse data - sPCA',
          legend=TRUE)

  #' }
}
\seealso{
\code{\link{pca}}, \code{\link{ipca}}, \code{\link{selectVar}},
\code{\link{plotIndiv}}, \code{\link{plotVar}} and http://www.mixOmics.org
for more details.
}
\author{
Ignacio Gonzalez, Kim-Anh Lê Cao, Fangzhou Yao, Al J Abadi
}
\keyword{algebra}
