% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-srbct.r
\docType{data}
\name{srbct}
\alias{srbct}
\title{Small version of the small round blue cell tumors of childhood data}
\format{
A list containing the following components: \describe{
\item{list("gene")}{data frame with 63 rows and 2308 columns. The expression
measure of 2308 genes for the 63 subjects.} \item{list("class")}{A class
vector containing the class tumour of each case (4 classes in total).}
\item{list("gene.name")}{data frame with 2308 rows and 2 columns containing
further information on the genes.} }
}
\source{
\url{https://www.research.nhgri.nih.gov/projects/Microarray/Supplement/index.html}
}
\usage{
data(srbct)
}
\value{
none
}
\description{
This data set from Khan \emph{et al.}, (2001) gives the expression measure
of 2308 genes measured on 63 samples.
}
\references{
Khan et al. (2001). Classification and diagnostic prediction of
cancers using gene expression profiling and artificial neural networks.
\emph{Nature Medicine} \bold{7}, Number 6, June.
}
\keyword{datasets}
