% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpca.R
\name{tpca}
\alias{tpca}
\title{Tensor PCA-like dimensionality reduction}
\usage{
tpca(
  x,
  ncomp = NULL,
  m = NULL,
  minv = NULL,
  center = TRUE,
  matrix_output = TRUE,
  bpparam = NULL
)
}
\arguments{
\item{x}{Tensor input.}

\item{ncomp}{The estimated number of components. ncomp can be explicitly set
using an integer value, 0 < float value < 1, or left as NULL which will
default to the maximum ncomp value possible.}

\item{m}{A function which applies an orthogonal tensor tubal transform.}

\item{minv}{The inverse of m.}

\item{center}{If set to false, the data tensor will not be centralized into
Mean Deviation Form (see Mor et al. 2022). By default, the mean horizontal
slice of the input tensor(s) are subtracted, so that all of the horizontal
slices sum to 0, analgous to centering matrix data.}

\item{matrix_output}{Collect the top singular vectors across the tensor,
organized by the magnitude of the corresponding singular vector, and place
them into the columns of a matrix. Corresponds to the 'matrix compression'
type of scheme described in Mor et al. 2022.}

\item{bpparam}{A \linkS4class{BiocParallelParam} object indicating the type
of parallelisation. Does not have any effect if transform functions
explicitly set using \code{m}, \code{minv}.}
}
\description{
Tensor analogue of PCA introduced by Mor et al. (2022) based on Kilmer's
m-product algebra and tsvdm.
}
\author{
Brendan Lu
}
