% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tpls.R
\name{tpls}
\alias{tpls}
\title{Tensor PLS-like regression}
\usage{
tpls(
  x,
  y,
  ncomp = NULL,
  m = NULL,
  minv = NULL,
  mode = "regression",
  center = TRUE,
  matrix_output = TRUE,
  bpparam = NULL
)
}
\description{
Developed @ Melbourne Integrative Genomics
}
\author{
Brendan Lu
}
