% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tens.tplsda.R
\name{tplsda}
\alias{tplsda}
\title{Run tensor PLSDA-like analysis}
\usage{
tplsda(
  x,
  y,
  multilevel = FALSE,
  ncomp = NULL,
  m = NULL,
  minv = NULL,
  center = TRUE,
  bpparam = NULL
)
}
\description{
Developed @ Melbourne Integrative Genomics
}
\author{
Brendan Lu
}
